requirements.txt
alembic==1.4.1
altgraph==0.17
antiorm==1.2.1
apispec==1.3.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.5
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
chatbot==1.5.2b0
chatterbot==1.1.0
chatterbot-corpus==1.2.0
cipher==0.1
Click==7.0
colorama==0.4.3
comtypes==1.1.7
crypto==1.4.1
cycler==0.10.0
DateTime==4.3
db==0.1.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-phone-field==1.8.0
django-registration-redux==2.8
django-session-timeout==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
Flask==1.1.1
Flask-AppBuilder==2.2.4
Flask-Babel==1.0.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==1.22.0
getmac==0.8.2
googlemaps==4.4.1
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.9
inflect==4.1.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jaraco.itertools==5.0.0
jedi==0.17.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
mathparse==0.1.2
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
MouseInfo==0.1.3
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.3
pafy==0.5.5
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
paytm-pg==1.0.6
pefile==2019.4.18
phonenumbers==8.12.3
pickleshare==0.7.5
Pillow==7.1.2
Pint==0.11
playsound==1.2.2
prison==0.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2==2.8.4
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycryptodome==3.9.7
pygame==1.9.6
PyGetWindow==0.0.8
Pygments==2.6.1
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.4.4
PyMsgBox==1.0.7
pyparsing==2.4.7
pyperclip==1.8.0
pypi==2.1
pypng==0.0.20
PyQRCode==1.2.1
PyRect==0.1.4
pyrsistent==0.15.7
PyScreeze==0.1.26
pyspeedtest==1.2.7
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pyttsx3==2.87
pytube==9.6.0
pytube3==9.6.4
PyTweening==1.0.3
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==3.13
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
ratelim==0.1.6
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
shellescape==3.8.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
SpeechRecognition==3.8.1
speedtest-cli==2.1.2
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tk==0.1.0
tornado==6.0.4
tqdm==4.45.0
trainer==0.0.1
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
wolframalpha==4.0.0
wrapt==1.11.2
WTForms==2.2.1
xmltodict==0.12.0
youtube-dl==2020.6.16.1
zope.interface==4.7.1